perf(ml-kem): a narrower rank, and one sponge where the widths agree
Two self-contained wins the #23 profile ranked, neither of which changes a
byte the scheme produces. Measured on CPU at the production shape (B = 256,
k = 3), A/B'd inside one process so process noise cancels.

**`_compact`'s rank fits `int16`.** It counts acceptances among 560
candidates, so it cannot leave `[0, 560]` — and the unrolled binary search
compares that whole row against every one of 256 slots, which is the file's
hottest traffic. Halving its width halves that: 2.28 ms -> 1.23 ms (-46%),
output identical, HLO count unchanged. The compaction is ~9% of `SampleNTT`,
which is the single largest stage of a decapsulation on CPU.

**`encrypt` draws once when `eta1 == eta2`.** `y` takes nonces 0..k-1 and
`e_1`/`e_2` take k..2k; the counter is contiguous, so equal widths make the
two draws one PRF over one range. Each `prf` call is a separately unrolled
sponge, so the second call is a second whole program rather than a second
pass: -8104 HLO ops on the compiled `encrypt` (-13.9%), byte-identical
output, warm runtime a wash to a few percent better.

The branch is not a runtime cost — `eta1` and `eta2` are Python ints off the
parameter set, so it resolves while tracing and each set compiles to one
program.

**Why it is a branch rather than one merged call.** Merging across *unequal*
widths would need `PRF_2(s,b) == PRF_3(s,b)[:128]`. That holds — SHAKE is an
XOF and it is byte-equal for all three sets — but FIPS 203 nowhere states it,
and `hashes.py` states the opposite rule about itself: an XOF's output length
is part of which hash it is, not a request made of one. ML-KEM-512 is the one
set with `eta1 != eta2`, and it keeps its two calls rather than resting the
scheme's correctness on an identity the standard does not promise.

The ACVP vectors, including the negative ones, pass unchanged at all three
parameter sets.

+++ enc_frx/ml_kem/sampling.py
@@ -126,8 +126,11 @@ def _compact(candidates: Array) -> Array:
     lets the module docstring promise the same wrong array everywhere rather than
     whatever each backend does at the edge.
     """
-    rank = fnp.cumsum((candidates < np.int32(Q)).astype(np.int32), axis=-1)
-    slots = fnp.arange(N, dtype=np.int32)
+    # `int16` because `rank` counts acceptances among `CANDIDATES` of them and so
+    # cannot leave `[0, 560]`: the unrolled binary search below compares the whole
+    # rank row against every slot, and halving its width halves that traffic.
+    rank = fnp.cumsum((candidates < np.int32(Q)).astype(np.int16), axis=-1)
+    slots = fnp.arange(N, dtype=np.int16)
     picked = frx.vmap(
         lambda row: fnp.searchsorted(row, slots, side="right", method="scan_unrolled")
     )(rank)
